LogSegmentMetadataStore should only notify when the list of log segments is updated
Currently it notifies the listeners not only when there is a change but also when session expires. it would break the readahead loop and cause readers have to wait until it is able to connect to zookeeper again.

With this change, it would only notify when the list of log segments is updated. If it disconnects to zookeeper, the listener won't be notified and it would keep reading from the log segments it knows.

RB_ID=842998