Handle correctly merge commits.
In the case of pull request #3, there was a merge commit in the pull
request (https://api.github.com/repos/apache/incubator-distributedlog/commits/2197d49261785a370f22c026cc9fe02f55b7e77c)
which has no login key for the actor, and the script was unable to
handle this situation. Now we check for the presence of the key.

Author: Franck Cuny <fcuny@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

Closes #11 from franckcuny/fcuny/fix-dl-merge-script