Make the shell style more consistent.
The style was inconsistent, using sometimes 4 spaces or 2 spaces for
indentation; using 'then' on the same line or another line; some
variables were quoted correctly; etc.

The style is now consistent and it's easier to read the script.

Instead of blindly running `sysctl`, only do it if we're on Linux.