Remove more duplicated shell scripts.
The script to run the various DLog components (benchmark, service,
etc) can be simplified by sharing most of the code. The common code is
moved to 'scripts/common.sh', and the other scripts source that one to
set a few variables, load a few environment settings, and build the
path for the jars.