DL-117: Stream metadata store
This change is to abstract the zookeeper operations into a stream metadata store, so we can replace zookeeper with other metadata store easily.

So the metadata operations in distributedlog now are managed by 3 classes:

- LogMetadataStore : it is the namespace metadata store : it manages the location (uri) mapping for streams and handle namespace operations.
- LogStreamMetadataStore: it is the stream metadata store : it manages the metadata for a single stream, such as managing read/write lock, retriving/creating stream metadata, deleting metadata and such.
- LogSegmentMetadataStore: it is the segment metadata store : it manages the log segment metadata for individual log segment.

LogMetadataStore and LogSegmentMetadataStore are already there. This change focus on LogStreamMetadataStore

Changed:

* abstract all the zookeeper metadata operation in log handlers to LogStreamMetadataStore
* remove disabling max tx id santify check, as maxTxId update is part of the metadata update transaction

Not changed:

the name of ZKLogMetadataForReader and ZKLogMetadataForWriter are not changed. I will send out a change to rename these two classes as they are not related to zookeeper anymore.