Do not assume git is available in the shell scripts.
Almost all the shell scripts were using `git` to figure out the root
directory of distributedlog, but we can't assume it is available on the
host. We could also not have a .git directory if the package is shipped
in a tarball or other package format.

Reformat some of the scripts to use consistent indentation level.