DL-206: Delete the log should also delete the underline ledgers
Problem:
We're not deleting the ledgers when we delete the whole log stream using dlm.delete() API. This would cause a lot of garbage/orphan ledgers in Bookkeeper.

The fix is to delete the ledger when we delete the log stream. Also added a test to validate.

Author: Yiming Zang <yzang@twitter.com>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #152 from yzang/yzang/DL-206