ISSUE #177: mvn release failed due to wrong scm url
Descriptions of the changes in this PR:

Reason:

we are using `gitbox` url as the scm connection. However gitbox is a readonly mirror of github repo.
The committers don't have permissions to commit directly to gitbox repo.

Solution:

update the scm connection to github url

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #178 from sijie/fix_scm_url, closes #177