ISSUE #222: post-commit ci job is broken
Descriptions of the changes in this PR:

The reason that the post commit ci job is broken is because `git remote show apache` is used in `publish-website.sh` script, but there isn't `apache` branch in CI cloned workspace. we need to change this from `apache` to `origin`.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #223 from sijie/debug_postcommit_job, closes #222