Include `nc` command in distributedlog docker image
Descriptions of the changes in this PR:

`nc` is usually used for checking if zookeeper is up and also verifying connectivity. It is good to have `nc` installed in the docker image by default. so a k8s helm script can use that for checking connectivities.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #241 from sijie/sijie/docker_image_nc